+++ packages/alice/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 * [BREAKING_CHANGE] Removed maxCallsCount. To change call count, use `storage` constructor parameter
   with call count.
+* [BREAKING_CHANGE] Replaced configuration parameter of Alice with `AliceConfiguration`.
+* Fixed issue with invalid count of calls in notification.
 
 # 1.0.0-dev.9